CR1 issues fixed

+++ szalkszop/szalkszop/Migrations/201807301853160_AddingCategoryList.cs
@@ -9,7 +9,7 @@ namespace szalkszop.Migrations
         {
             AddColumn("dbo.Products", "Name", c => c.String());
             AddColumn("dbo.Products", "Product_Id", c => c.Int());
-            AddColumn("dbo.ProductCategories", "ProductCategory_Id", c => c.Byte());
+            AddColumn("dbo.ProductCategories", "ProductCategory_Id", c => c.Int());
             CreateIndex("dbo.Products", "Product_Id");
             CreateIndex("dbo.ProductCategories", "ProductCategory_Id");
             AddForeignKey("dbo.ProductCategories", "ProductCategory_Id", "dbo.ProductCategories", "Id");
